+++ src/pages/LoginPage.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Main from '../components/Main';
-import { LoginPageHeader, Login } from '../components/Login';
+import LoginPageHeader from '../components/LoginHeader';
+import Login from '../components/Login';
 
 function LoginPage() {
   return (
